requirements.txt
langchain>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
wikipedia>=1.4.0
google-search-results>=2.4.2
python-jose>=3.3.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0 